reorganize txm
