+++ relayer/ops/test_helpers.go
@@ -1,4 +1,4 @@
-package txm
+package ops
 
 import (
 	"bytes"
