+++ relayer/pkg/chainlink/config/config.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink-relay/pkg/logger"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/db"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/txm"
 )
 
 var DefaultConfigSet = ConfigSet{
@@ -31,12 +32,14 @@ type ConfigSet struct {
 }
 
 type Config interface {
+	txm.TxConfig // txm config
+
+	// ocr2 config
 	OCR2CachePollPeriod() time.Duration
 	OCR2CacheTTL() time.Duration
+
+	// client config
 	RequestTimeout() time.Duration
-	TxTimeout() time.Duration
-	TxSendFrequency() time.Duration
-	TxMaxBatchSize() int
 
 	Update(db.ChainCfg)
 }
